+++ R/methods_lme4.R
@@ -64,8 +64,8 @@ get_predict.lmerMod <- get_predict.merMod
 
 #' @rdname sanitize_model_specific
 #' @export
-sanitize_model_specific.merMod <- function(model, re.form, ...) {
-    if (missing(re.form) || (!isTRUE(is.na(re.form)))) {
+sanitize_model_specific.merMod <- function(model, re.form, vcov = TRUE, ...) {
+    if (!isFALSE(vcov) && (missing(re.form) || (!isTRUE(is.na(re.form))))) {
         msg <- "For this model type, `marginaleffects` only takes into account the uncertainty in fixed-effect parameters. This is often appropriate when `re.form=NA`, but may be surprising to users who set `re.form=NULL` (default) or to some other value. Call `options(marginaleffects_safe = FALSE)` to silence this warning."
         warn_sprintf(msg)
     }